+++ Soggins/ExternalAPIs/IMoreDifficultiesApi.cs
@@ -1,0 +1,6 @@
+﻿namespace Shockah.Soggins;
+
+public interface IMoreDifficultiesApi
+{
+	void RegisterAltStarters(Deck deck, StarterDeck starterDeck);
+}